requirements.txt
Pandas
matplotlib
numpy
scikit-learn
seaborn
tensorflow
dvc
notebook
scipy
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-pyYAML
Flask
Flask-cors
-e .